Full upgrade + core prior to repo-switch
As we have to rewrite composer.lock the simplest way of switching the composer-repo is to do a full upgrade via the old repo, then remove composer.lock and doing a composer install. For this we need everything to be as upgraded as possible.

As a part of the upgrade I had to switch to a newer version of the metatag patch.

DDSDK-337