Disabling core search and re-exporting configuration
This is done to fix an issue where core-search would index newly created events without theming during batch-runs. The root of the problem is that we choose to use cores view-mode for our own styling, but introducing a new view-mode would be a lot more work than just simply disabling a module we don't use (core search).

DDSDK-364