Readd search block to header
When disabling the Core Search module the search form in the header 
also disappears. This is because the search form is a part of the 
moduloe.

We do not want to enable Core Search so instead we recreate the parts
we are missing:

- The form from which searches are executed
- The block which places the form on the page

We also add the new block to the header to replace the obsolete search
block.

We have to rename a template to take the new block name into account.


DDSDK-370