Make composer fail if patches do not apply
By default the composer-patches will print out a message if a patch does not apply but the process continues. We want the process to fail entirely to ensure that patch failures hare handled properly.

The other option would be to rely on the COMPOSER_EXIT_ON_PATCH_FAILURE environment variable but we want to ensure that this happens across environments.