Use interfaces when checking `instanceof`
Specifically because `Node` no longer inherits from `Entity` (see
https://git.drupalcode.org/project/drupal/commit/47dab9a4a9b7c606cab4a1b67d68a316d035eb2f).

And while we are at it cleanup some other places with regards to the
same.

Also fix `$term->getVocabularyId()` which has been deprecated in favor
of `bundle()`.

And while we are at it do the same for `$node->getType()` (where we
change anyway ) although it haven't been deprecated (yet?).

See DDSDK-438.