Update Drupal Core to 8.9.1
This addresses SA-CORE-2020-004, SA-CORE-2020-005 and 
SA-CORE-2020-006. It also updates the project to the latest Core 
minor version before 9.0.

We also need Drupal Scaffold in that version to avoid problems with
file permissions in web/sites/default in for release 8.8.x.

Also reraise requirements for drupal/pathauto. It was mistakenly
Lowered in the previous version.

DDSDK-462