Fix issues with index.php in URLs. DDSDK-479
There is an issue, where index.php 'randomly' appears in
some urls. An example:
https://dds.dk/index.php/artikel/nye-spejderchefer-vi-skal-skabe-modige-born-og-unge

It's not all pages, and there is not a clear pattern why
it happens sometimes. The 'real' path is valid, but the
issue appears when links on the website actually links
to the index.php versions.

It appears to be an obscure bug in D8 when using
platform.sh and sitemap/pathauto/others(?)

This commit implements a quickfix, described here:
https://www.drupal.org/project/drupal/issues/3050261

'Just a quick note to let folks know that we installed the
Redirect module and enabled Enforce clean and canonical URLs,
and haven't had the problem return since (it's been weeks).'

It appears to work for DDS too.