Add new hero-banner paragraph - DDSDK-741
Completely reworked/redesigned hero-banner that removes the link situations (they will live in another paragraph)
and changes title text style, image position and view mode.
We also render the search form block inside the hero-banner along a link list of popular search words.
The search words are added manually by the editor in the hero-banner paragraph form as clean strings.
The strings are being transformed to links in a preprocess to keep logic in the twig template to a minimum.