Redesign situation links paragraph - DDSDK-738
Implements a completely new design and focus on not breaking the track
version of this paragraph. Before the track styling was depended on the
default version css but they are now so different that I've made them
completely independant.

The styling, naming and HTML semantics are all over the place but I didn't want to refactor
too much code.