Redesign/restyle shortcuts links paragraph - DDSDK-745
Removed the title, changed background color and link icon and text.

The deck paragraph css was controlling the padding and background color
of the shortcuts links so I had to use important to override these rules.
I wanted to disable the rule but I'm afraid it could break a lot of things.

Looks like the content-list also relies on it to get background color
but it was hard to test since the column deck 2 paragraph is broken when displaying
the content-list. It's buggy and out of this tasks scope. Hopefully we have
time to fix some of these bugs later on.