Change the global body background color - DDSDK-751
The body background color has been changed to a light sand color.
The grey paragraph background color has been changed to a sand color.
I tried to change the name of the grey color to match the actual color applied.
The value comes from a list field where the value is in use, so we can't change it.
In the future we want to solve this issue.